+++ Sources/XcodeSupport/XcodeProjectlike.swift
@@ -4,12 +4,13 @@ import PeripheryKit
 
 protocol XcodeProjectlike: AnyObject {
     var path: FilePath { get }
-    var targets: Set<XcodeTarget> { get } // Set to ensure uniqueness
+    var targets: Set<XcodeTarget> { get }
+    var packageTargets: [SPM.Package: Set<SPM.Target>] { get }
     var type: String { get }
     var name: String { get }
     var sourceRoot: FilePath { get }
 
-    func schemes() throws -> Set<XcodeScheme> // Set to ensure uniqueness
+    func schemes() throws -> Set<XcodeScheme>
 }
 
 extension XcodeProjectlike {