+++ Tests/XcodeTests/UIKitProject/LocalPackages/LocalPackage/Sources/LocalPackageTarget/LocalPackageUsedType.swift
@@ -1,0 +1,3 @@
+public struct LocalPackageUsedType {
+    public init() {}
+}