+++ Tests/XcodeTests/UIKitProject/LocalPackages/LocalPackage/Tests/LocalPackageTargetTests/LocalPackageTargetTests.swift
@@ -1,0 +1,8 @@
+import XCTest
+@testable import LocalPackageTarget
+
+final class LocalPackageTargetTests: XCTestCase {
+    func testExample() throws {
+        _ = LocalPackageUsedInTestType()
+    }
+}