+++ Tests/XcodeTests/UIKitProject/UIKitProject/AppDelegate.swift
@@ -1,9 +1,11 @@
 import UIKit
+import LocalPackageTarget
 
 @UIApplicationMain
 class AppDelegate: UIResponder, UIApplicationDelegate {
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
         // Override point for customization after application launch.
+        _ = LocalPackageUsedType()
         return true
     }
 
@@ -20,6 +22,4 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         // If any sessions were discarded while the application was not running, this will be called shortly after application:didFinishLaunchingWithOptions.
         // Use this method to release any resources that were specific to the discarded scenes, as they will not return.
     }
-
-
 }